* thread_pthread.c (native_sleep): suppress a warning for platforms   which time_t is int.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@29052 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ thread_pthread.c
@@ -612,7 +612,7 @@ native_sleep(rb_thread_t *th, struct timeval *tv)
 	}
     }
 
-    thread_debug("native_sleep %ld\n", tv ? tv->tv_sec : -1);
+    thread_debug("native_sleep %ld\n", (long)(tv ? tv->tv_sec : -1));
     GVL_UNLOCK_BEGIN();
     {
 	pthread_mutex_lock(&th->interrupt_lock);
